+++ CHANGELOG.rst
@@ -43,3 +43,4 @@ latest
 ------
 
 * Add https://pre-commit.com configuration.
+* Use find_shortest_chains instead of find_shortest_chain on the Grimp import graph.